% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Linear_MasterSample.R
\name{linear_masterSample}
\alias{linear_masterSample}
\title{Generate sample points on linear features in New Zealand using a HIP based master sample}
\usage{
linear_masterSample(island = "North", shp, stratum, oversamples = 1, n = 10)
}
\description{
Generates HIP sample points in a specified sample frame based on New Zealand linear features mastersample. Users need to specify 'island' the sample frame is in and how many points are required.
}
